report loss

+++ app.js
@@ -56,7 +56,9 @@ bot.onText(/\/register/, function(msg, match) {
     var user = {first_name: msg.from.first_name,
             last_name: msg.from.last_name || "",
             username: msg.from.username.toLowerCase(),
-            challenge: false
+            challenge: false,
+            wins: 0,
+            loses: 0
           };
     db.saveUser(user, chatId).then(function(){
       bot.sendMessage(chatId, "You're all set. Challenge someone by typing /challenge [name]");
@@ -95,26 +97,34 @@ bot.onText(/\/challenge/, function(msg, match) {
 bot.onText(/\/accept/, function(msg) {
   console.log('MSG accept->',msg);
   var chatId = msg.chat.id;
-
-  db.getChallenge(chatId).then(function(challenge){
-    console.log('Accept challenge:', challenge);
-    var match = {};
-    match.host    = challenge.userFromId;
-    match.visitor = chatId;
-    match.result  = false;
-    match.winner  = false;
-    db.saveMatch(match).then(function(){
-      db.removeChallenge(chatId);
-      bot.sendMessage(chatId, "Challenge accepted. Let's Play.");
-      bot.sendMessage(challenge.userFromId, "Your challenge to @"+msg.chat.username+" has been accepted. Get ready!");
-    }, function() {
-      console.log('Error saving the match.');
-      bot.sendMessage(chatId, "There was an error creating your match.");
-    });
-  }, function(){
-    console.log('Error accepting challenge');
-    bot.sendMessage(chatId, "You don't have any challenge to accept.");
+  db.hasPendingMatch(chatId).then(function(hasPendingMatch){
+    if (!hasPendingMatch) {
+      db.getChallenge(chatId).then(function(challenge){
+        console.log('Accept challenge:', challenge);
+        var match = {};
+        match.host    = challenge.userFromId;
+        match.visitor = chatId;
+        match.result  = false;
+        match.winner  = false;
+        db.saveMatch(match).then(function(){
+          db.removeChallenge(chatId);
+          bot.sendMessage(chatId, "Challenge accepted. Let's Play.");
+          bot.sendMessage(challenge.userFromId, "Your challenge to @"+msg.chat.username+" has been accepted. Get ready!");
+        }, function() {
+          console.log('Error saving the match.');
+          bot.sendMessage(chatId, "There was an error creating your match.");
+        });
+      }, function(){
+        console.log('Error accepting challenge');
+        bot.sendMessage(chatId, "You don't have any challenge to accept.");
+      });
+    } else {
+      bot.sendMessage(chatId, "You cannot accept a challenge because you have an unfinished match. Report the result or resign the match and try again.");
+    }
+  }, function() {
+    console.log('Error checking if hasPendingMatch');
   });
+
 });
 
 //decline a challenge
@@ -137,16 +147,27 @@ bot.onText(/\/decline/, function(msg) {
 });
 
 //Report a match
-bot.onText(/\/report/, function(msg) {
-  console.log('MSG report->',msg);
+bot.onText(/\/lost/, function(msg) {
+  console.log('MSG lost->',msg);
   var chatId = msg.chat.id,
-  userFromId = msg.from.id,
-  arg = msg.text.split('/report ')[1],
-  args = arg.split(' '),
-  userTo = args[0].split('@')[1],
-  score = args[1],
-  sets = score.split(","),
-  matchesRef = fireRef.child("matches");
+  loserId = msg.from.id,
+  arg = msg.text.split('/lost ')[1],
+  sets = arg.split(' ');
 
-  console.log('userTo->',userTo, 'score->',score);
+  //TODO: See if this is not double check.
+  db.hasPendingMatch(chatId).then(function(hasPendingMatch){
+    if (hasPendingMatch) {
+      db.findMatchWithoutResultByPlayerId(chatId).then(function(matchKey, match) {
+        db.updateMatchResult(matchKey, match, sets, loserId);
+      }, function() {
+        console.log('Error checking if the player has match without result');
+      });
+    } else {
+      console.log('Player does not have a match without result.');
+      bot.sendMessage(chatId, "You don't have any match to report.");
+    }
+  }, function() {
+    console.log('Error checking if hasPendingMatch');
+  });
+
 });
